fix: Ads take long to init in offline mode, other minor fixes (#44)

+++ lib/core/backgrounds/domain/use_case/get_available_backgrounds_use_case.dart
@@ -27,16 +27,18 @@ class GetAvailableBackgroundsUseCase {
     final failureOrSuccess = await _backgroundsRepo.getBackgroundsRemote(
       languageCode: languageCode,
     );
-    final entries = failureOrSuccess.getEntries();
-    final activeBackground = entries.$2?.$1.toNullable();
-    final availableBackgrounds = entries.$2?.$2 ?? const <AvailableBackground>[];
+    final (failure, (tuple)) = failureOrSuccess.getEntries();
+    final activeBackground = tuple?.$1.toNullable();
+    final availableBackgrounds = tuple?.$2 ?? const <AvailableBackground>[];
     
-    if (activeBackground != null) {
-      unawaited(_activeBackgroundCubit.setCustomBackground(activeBackground));
-      unawaited(_backgroundsRepo.storeBackgroundAssetLocal(activeBackground));
-    } else {
-      unawaited(_activeBackgroundCubit.clearState());
-      unawaited(_backgroundsRepo.deleteBackgroundAssetLocal());
+    if (failure == null) {
+      if (activeBackground != null) {
+        unawaited(_activeBackgroundCubit.setCustomBackground(activeBackground));
+        unawaited(_backgroundsRepo.storeBackgroundAssetLocal(activeBackground));
+      } else {
+        unawaited(_activeBackgroundCubit.clearState());
+        unawaited(_backgroundsRepo.deleteBackgroundAssetLocal());
+      }
     }
 
     if (availableBackgrounds.isNotEmpty) {
